fix(workflows): load model options via server route
